// apps/web/src/http/gen/model/get-telemetry-fields-any-of-item.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * LAPES - API
 * API for supervisory control and data acquisition
 * OpenAPI spec version: 1.0.0
 */

export type GetTelemetryFieldsAnyOfItem =
  (typeof GetTelemetryFieldsAnyOfItem)[keyof typeof GetTelemetryFieldsAnyOfItem]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetTelemetryFieldsAnyOfItem = {
  anguloFaseA: "anguloFaseA",
  anguloFaseB: "anguloFaseB",
  anguloFaseC: "anguloFaseC",
  correnteA: "correnteA",
  correnteB: "correnteB",
  correnteC: "correnteC",
  correnteNeutroCalculado: "correnteNeutroCalculado",
  correnteNeutroMedido: "correnteNeutroMedido",
  fpDeslocamentoFaseA: "fpDeslocamentoFaseA",
  fpDeslocamentoFaseB: "fpDeslocamentoFaseB",
  fpDeslocamentoFaseC: "fpDeslocamentoFaseC",
  fpDeslocamentoTotal: "fpDeslocamentoTotal",
  fpRealFaseA: "fpRealFaseA",
  fpRealFaseB: "fpRealFaseB",
  fpRealFaseC: "fpRealFaseC",
  fpRealTotalAritmetica: "fpRealTotalAritmetica",
  fpRealTotalVetorial: "fpRealTotalVetorial",
  frequencia: "frequencia",
  phiFaseA: "phiFaseA",
  phiFaseB: "phiFaseB",
  phiFaseC: "phiFaseC",
  potenciaAparenteA: "potenciaAparenteA",
  potenciaAparenteB: "potenciaAparenteB",
  potenciaAparenteC: "potenciaAparenteC",
  potenciaAparenteTotalAritmetica: "potenciaAparenteTotalAritmetica",
  potenciaAparenteTotalVetorial: "potenciaAparenteTotalVetorial",
  potenciaAtivaFundamentalA: "potenciaAtivaFundamentalA",
  potenciaAtivaFundamentalB: "potenciaAtivaFundamentalB",
  potenciaAtivaFundamentalC: "potenciaAtivaFundamentalC",
  potenciaAtivaFundamentalHarmonicaA: "potenciaAtivaFundamentalHarmonicaA",
  potenciaAtivaFundamentalHarmonicaB: "potenciaAtivaFundamentalHarmonicaB",
  potenciaAtivaFundamentalHarmonicaC: "potenciaAtivaFundamentalHarmonicaC",
  potenciaAtivaFundamentalHarmonicaTotal:
    "potenciaAtivaFundamentalHarmonicaTotal",
  potenciaAtivaFundamentalTotal: "potenciaAtivaFundamentalTotal",
  potenciaAtivaHarmonicaA: "potenciaAtivaHarmonicaA",
  potenciaAtivaHarmonicaB: "potenciaAtivaHarmonicaB",
  potenciaAtivaHarmonicaC: "potenciaAtivaHarmonicaC",
  potenciaAtivaHarmonicaTotal: "potenciaAtivaHarmonicaTotal",
  potenciaReativaA: "potenciaReativaA",
  potenciaReativaB: "potenciaReativaB",
  potenciaReativaC: "potenciaReativaC",
  potenciaReativaTotalAritmetica: "potenciaReativaTotalAritmetica",
  potenciaReativaTotalVetorial: "potenciaReativaTotalVetorial",
  temperaturaSensorInterno: "temperaturaSensorInterno",
  tensaoFaseFaseAB: "tensaoFaseFaseAB",
  tensaoFaseFaseBC: "tensaoFaseFaseBC",
  tensaoFaseFaseCA: "tensaoFaseFaseCA",
  tensaoFaseNeutroA: "tensaoFaseNeutroA",
  tensaoFaseNeutroB: "tensaoFaseNeutroB",
  tensaoFaseNeutroC: "tensaoFaseNeutroC",
  thdCorrenteA: "thdCorrenteA",
  thdCorrenteB: "thdCorrenteB",
  thdCorrenteC: "thdCorrenteC",
  thdTensaoA: "thdTensaoA",
  thdTensaoB: "thdTensaoB",
  thdTensaoC: "thdTensaoC",
} as const